# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agustinos.tv API",
        "version": "3.20.01"
    },
    "servers": [
        {
            "url": "http://api.agustinos.tv"
        }
    ],
    "paths": {
...
        "/api/auth/activate/{token}": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Activar una cuenta",
                "description": "Activar una cuenta",
                "operationId": "App\\Http\\Controllers\\AuthController::Activar",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "description": "Email de Usuario a autenticar",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
...
        "/api/canales": {
            "get": {
                "tags": [
                    "Canales"
                ],
                "summary": "Listar canales",
                "description": "Obtener la lista de canales",
                "operationId": "App\\Http\\Controllers\\CanalesController::index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de Página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista resultado de los canales"
                    }
                }
            },
            "post": {
                "tags": [
                    "Canales"
                ],
                "summary": "Guardar un canal",
                "description": "Guardar un canal",
                "operationId": "App\\Http\\Controllers\\CanalesController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre de el canal",
                                        "type": "string"
                                    },
                                    "estado": {
                                        "description": "Estado del canal ['ONLINE','OFFLINE']",
                                        "type": "string"
                                    },
                                    "ruta": {
                                        "description": "Ruta de la emisión",
                                        "type": "string"
                                    },
                                    "delegacion_id": {
                                        "description": "Id de la delegación",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Canal no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Canales"
                ],
                "summary": "Eliminar un canal",
                "description": "Eliminar un canal",
                "operationId": "App\\Http\\Controllers\\CanalesController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Canal no encontrado"
                    }
                }
            }
        },
...
        "/api/users/": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar usuarios",
                "description": "Obtener un usuario",
                "operationId": "App\\Http\\Controllers\\UserController::index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de Página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista resultado de los usuarios"
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Guardar un usuario",
                "description": "Guardar un usuario",
                "operationId": "App\\Http\\Controllers\\UserController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "Nombre de usuario",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email de usuario",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Contraseña del usuario",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "409": {
                        "description": "El email introducido ya existe en el sitema"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar un usuario",
                "description": "Eliminar un usuario",
                "operationId": "App\\Http\\Controllers\\UserController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "403": {
                        "description": "No está permitido eliminar el usuario que está actualmente logueado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        },
...
    }
}